Update clock versions (order)
