+++ functions/clock1.py
@@ -11,94 +11,69 @@ import logger
 from control.led_strip import set_brightness_depending_on_daytime
 from functions.effects import clear
 
-# color hours
-hR = 100
-hG = 20
+hR = 200
+hG = 0
 hB = 0
 
-# color minutes
-mR = 20
+mR = 0
 mG = 0
-mB = 100
+mB = 200
 
-# color seconds
-sR = 6
-sG = 30
-sB = 10
+sR = 92
+sG = 67
+sB = 6
 
 name = "Clock 1"
 log = logger.get_logger(name)
 
 
-def run_clock1(stripe):
+def run_clock1(strip):
     from control import get_stop_flag
     while not get_stop_flag():
+        # noinspection PyBroadException
         try:
+            now = set_brightness_depending_on_daytime(strip)[0]
+            hour = int(int(now.hour) % 12 * 2)
+            minute = int(now.minute // 2.5)
+            second = int(now.second // 2.5)
 
-            now = set_brightness_depending_on_daytime(stripe)[0]
-            led_for_hour = int(int(now.hour) % 12 * 2)
-            led_for_minute = int(now.minute // 2.5)
-            leds_per_2500ms = int(round(now.second / 2.5))
+            for i in range(0, strip.numPixels(), 1):
+                # hour
+                strip.setPixelColorRGB(hour, hG, hR, hB)
 
-            _dial(stripe)
-            _seconds(leds_per_2500ms, stripe)
-            _minute(led_for_minute, led_for_hour, stripe)
-            _hour(led_for_hour, stripe)
+                # minute
+                if minute == hour:
+                    if 12 < minute < strip.numPixels():
+                        if hour <= 23:
+                            strip.setPixelColorRGB(hour + 1, hG, hR, hB)
+                            strip.setPixelColorRGB(minute, mG, mR, mB)
+                        else:
+                            strip.setPixelColorRGB(0, hG, hR, hB)
+                            strip.setPixelColorRGB(minute - 1, mG, mR, mB)
+                    else:
+                        strip.setPixelColorRGB(minute + 1, mG, mR, mB)
+                else:
+                    strip.setPixelColorRGB(minute, mG, mR, mB)
 
-            stripe.show()
-            if leds_per_2500ms == stripe.numPixels():
-                time.sleep(1.5)
-                clear(stripe)
+                # second
+                if i == second:
+                    strip.setPixelColorRGB(i, sG, sR, sB)
+                else:
+                    strip.setPixelColorRGB(i, 0, 0, 0)
+
+            strip.show()
+            time.sleep(0.1)
 
         except KeyboardInterrupt:
+            print()
             log.warn("KeyboardInterrupt.")
             exit()
 
         except Exception as e:
             log.error("Any error occurs: " + str(e))
             exit()
-    clear(stripe)
 
-
-def _seconds(leds_per_2500ms, stripe):
-    for led in range(0, leds_per_2500ms, 1):
-        if 0 < (led + 1) < stripe.numPixels():
-            stripe.setPixelColorRGB(led + 1, sG, sR, sB)
-        if (led + 1) == stripe.numPixels():
-            stripe.setPixelColorRGB(0, sG, sR, sB)
-
-
-def _minute(led, led_hour, stripe):
-    if led < stripe.numPixels():
-        if led == led_hour:
-            _set_minute_led_before_and_after(stripe, led)
-        else:
-            stripe.setPixelColorRGB(led, mG, mR, mB)
-    if led >= stripe.numPixels():
-        if led == led_hour:
-            _set_minute_led_before_and_after(stripe, led_hour)
-            stripe.setPixelColorRGB(0, mG, mR, mB)
-        else:
-            stripe.setPixelColorRGB(0, mG, mR, mB)
-    else:
-        stripe.setPixelColorRGB(led, mG, mR, mB)
-
-
-def _set_minute_led_before_and_after(stripe, led):
-    stripe.setPixelColorRGB(led - 1, (mG // 5), (mR // 5), (mB // 5))
-    stripe.setPixelColorRGB(led + 1, (mG // 5), (mR // 5), (mB // 5))
-
-
-def _hour(led, stripe):
-    stripe.setPixelColorRGB(led, hG, hR, hB)
-
-
-def _dial(stripe):
-    dial = [0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22]  # hours
-    # dial = [0, 6, 12, 18]  # quarter only
-    for led in dial:
-        stripe.setPixelColorRGB(led, 125 // 10, 195 // 10, 30 // 10)  # warm yellow
-        # stripe.setPixelColorRGB(led, 15, 15, 15)  # white
+    clear(strip)
 
 
 if __name__ == '__main__':
