+++ functions/clock2.py
@@ -1,11 +1,14 @@
 #!/usr/bin/python3
 # -*- coding: utf-8 -*-
+
 __author___ = "Thomas Kaulke"
 __email__ = "kaulketh@gmail.com"
 __maintainer___ = "Thomas Kaulke"
 __status__ = "Development"
 
 import time
+
+from neopixel import Color
 
 import logger
 from control.led_strip import set_brightness_depending_on_daytime
@@ -19,10 +22,6 @@ mR = 0
 mG = 0
 mB = 200
 
-sR = 92
-sG = 67
-sB = 6
-
 name = "Clock 2"
 log = logger.get_logger(name)
 
@@ -30,17 +29,18 @@ log = logger.get_logger(name)
 def run_clock2(strip):
     from control import get_stop_flag
     while not get_stop_flag():
-        # noinspection PyBroadException
         try:
             now = set_brightness_depending_on_daytime(strip)[0]
             hour = int(int(now.hour) % 12 * 2)
             minute = int(now.minute // 2.5)
-            second = int(now.second // 2.5)
 
             for i in range(0, strip.numPixels(), 1):
                 # hour
-                strip.setPixelColorRGB(hour, hG, hR, hB)
-
+                if 12 < minute <= 23:
+                    strip.setPixelColorRGB(hour, hG, hR, hB)
+                    strip.setPixelColorRGB(hour + 1, hG // 4, hR // 4, hB // 4)
+                else:
+                    strip.setPixelColorRGB(hour, hG, hR, hB)
                 # minute
                 if minute == hour:
                     if 12 < minute < strip.numPixels():
@@ -54,26 +54,39 @@ def run_clock2(strip):
                         strip.setPixelColorRGB(minute + 1, mG, mR, mB)
                 else:
                     strip.setPixelColorRGB(minute, mG, mR, mB)
-
-                # second
-                if i == second:
-                    strip.setPixelColorRGB(i, sG, sR, sB)
-                else:
-                    strip.setPixelColorRGB(i, 0, 0, 0)
-
             strip.show()
-            time.sleep(0.1)
-
+            time.sleep(150)
+            _wipe_second(strip, (mG // 5, mR // 5, mB // 5), minute, backward=True)
+            clear(strip)
         except KeyboardInterrupt:
             print()
             log.warn("KeyboardInterrupt.")
             exit()
-
         except Exception as e:
             log.error("Any error occurs: " + str(e))
             exit()
-
     clear(strip)
+
+
+def _wipe_second(stripe, color, begin=0, backward=False):
+    if backward:
+        wait_ms = ((1000.0 // stripe.numPixels()) // 2) / 1000.0
+    else:
+        wait_ms = (1000.0 // stripe.numPixels()) / 1000.0
+
+    for i in range(begin + 1, stripe.numPixels() + begin):
+        if i >= stripe.numPixels():
+            i -= stripe.numPixels()
+        stripe.setPixelColor(i, Color(color[0], color[1], color[2]))
+        stripe.show()
+        time.sleep(wait_ms)
+    if backward:
+        for i in range(stripe.numPixels() + begin - 1, begin, -1):
+            if i >= stripe.numPixels():
+                i -= stripe.numPixels()
+            stripe.setPixelColor(i, Color(0, 0, 0))
+            stripe.show()
+            time.sleep(wait_ms)
 
 
 if __name__ == '__main__':
